server: refactor StartTenant() to pkg/server/tenant.go
Previously, StartTenant() function resided in pkg/server/testserver.go.
However, since this function is not solely used for testing purposes,
it is moved to its own file.

Release note: None
